+++ android/src/main/kotlin/com/adyen/checkout/flutter/utils/ConfigurationMapper.kt
@@ -146,6 +146,7 @@ object ConfigurationMapper {
             analyticsOptionsDTO = analyticsOptionsDTO,
             shopperLocale = shopperLocale,
             amount = amount,
+            threeDS2ConfigurationDTO = threeDS2ConfigurationDTO,
         )
 
     fun InstantPaymentConfigurationDTO.toCheckoutConfiguration(): CheckoutConfiguration =