+++ ios/adyen_checkout/Sources/adyen_checkout/components/action/ActionComponentManager.swift
@@ -20,8 +20,9 @@ class ActionComponentManager {
     func handleAction(actionComponentConfiguration: ActionComponentConfigurationDTO, componentId: String, actionResponse: [String?: Any?]) {
         do {
             let adyenContext = try actionComponentConfiguration.createAdyenContext()
-            actionComponent = CheckoutActionComponent(context: adyenContext)
-            actionComponent?.delegate = self
+            actionComponent = configuration.map {
+                AdyenActionComponent(context: adyenContext, configuration: $0)
+            } ?? AdyenActionComponent(context: adyenContext)            actionComponent?.delegate = self
             actionComponent?.presentationDelegate = getViewController()
             let jsonData = try JSONSerialization.data(withJSONObject: actionResponse, options: [])
             let action = try JSONDecoder().decode(Action.self, from: jsonData)